Scroll menu options when too many for the screen size
