+++ .gitignore
@@ -564,7 +564,10 @@ packages/app-mobile/components/CameraView/types.js
 packages/app-mobile/components/CameraView/utils/fitRectIntoBounds.js
 packages/app-mobile/components/CameraView/utils/useBarcodeScanner.js
 packages/app-mobile/components/Checkbox.js
-packages/app-mobile/components/DialogManager.js
+packages/app-mobile/components/DialogManager/PromptDialog.js
+packages/app-mobile/components/DialogManager/hooks/useDialogControl.js
+packages/app-mobile/components/DialogManager/index.js
+packages/app-mobile/components/DialogManager/types.js
 packages/app-mobile/components/DismissibleDialog.js
 packages/app-mobile/components/Dropdown.test.js
 packages/app-mobile/components/Dropdown.js
